requirements.txt
alembic==1.12.1
bcrypt==4.2.1
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
croniter==6.0.0
cryptography==44.0.0
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
Mako==1.2.4
MarkupSafe==2.1.5
onetimepass==1.0.1
paramiko==3.5.1
pycparser==2.22
PyMySQL==1.1.1
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
